Split the non gui related function out to a class of static functions. these functions can also be used from other places in the code.
